+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Changes to the plugin.
 
+## Unpublished
+
+### Changed
+
+- Outline now lists all #options in a *choice.
+
 ## [1.2.0] - 2020-03-05
 
 ### Added